+++ src/main/java/org/kdb/inside/brains/psi/AssignmentType.java
@@ -1,0 +1,9 @@
+package org.kdb.inside.brains.psi;
+
+public enum AssignmentType {
+    LOCAL,
+    GLOBAL,
+    PARAMETER,
+    TABLE_COLUMN,
+    QUERY_COLUMN
+}