+++ src/main/java/org/kdb/inside/brains/psi/QPsiUtil.java
@@ -1,12 +1,48 @@
-package org.kdb.inside.brains;
+package org.kdb.inside.brains.psi;
 
+import com.intellij.psi.PsiElement;
 import com.intellij.psi.util.PsiTreeUtil;
 import org.jetbrains.annotations.NotNull;
-import org.kdb.inside.brains.psi.*;
 
 import java.util.Optional;
 
 public class QPsiUtil {
+
+
+    public static AssignmentType getAssignmentType(QVariableElement variable) {
+        final PsiElement parent = variable.getParent();
+        if (parent instanceof QParameters) {
+            return AssignmentType.PARAMETER;
+        }
+
+        if (parent instanceof QTableColumn) {
+            return AssignmentType.TABLE_COLUMN;
+        }
+        if (parent instanceof QQueryColumn) {
+            return AssignmentType.QUERY_COLUMN;
+        }
+
+        if (!(parent instanceof QVariableAssignment)) {
+            return null;
+        }
+
+        final String qualifiedName = variable.getQualifiedName();
+        if (QVariableElement.hasNamespace(qualifiedName)) {
+            return AssignmentType.GLOBAL;
+        }
+
+        final QLambda lambda = variable.getContext(QLambda.class);
+        if (lambda == null) {
+            return AssignmentType.GLOBAL;
+        }
+
+        final PsiElement colon = PsiTreeUtil.findSiblingForward(variable, QTypes.COLON, true, null);
+        if (colon != null && colon.getNextSibling().getNode().getElementType() == QTypes.COLON) {
+            return AssignmentType.GLOBAL;
+        }
+        return AssignmentType.LOCAL;
+    }
+
     @NotNull
     public static String getDescriptiveName(@NotNull QVariable element) {
         final String fqnOrName = element.getQualifiedName();