+++ src/main/java/org/kdb/inside/brains/psi/refs/QReferenceContributor.java
@@ -4,7 +4,7 @@ import com.intellij.patterns.PlatformPatterns;
 import com.intellij.psi.PsiReferenceContributor;
 import com.intellij.psi.PsiReferenceRegistrar;
 import org.jetbrains.annotations.NotNull;
-import org.kdb.inside.brains.psi.QImportFile;
+import org.kdb.inside.brains.psi.QImportElement;
 import org.kdb.inside.brains.psi.QSymbol;
 import org.kdb.inside.brains.psi.QVariable;
 
@@ -13,6 +13,6 @@ public class QReferenceContributor extends PsiReferenceContributor {
     public void registerReferenceProviders(@NotNull PsiReferenceRegistrar registrar) {
         registrar.registerReferenceProvider(PlatformPatterns.psiElement(QSymbol.class), new QSymbolReferenceProvider());
         registrar.registerReferenceProvider(PlatformPatterns.psiElement(QVariable.class), new QVariableReferenceProvider());
-        registrar.registerReferenceProvider(PlatformPatterns.psiElement(QImportFile.class), new QImportReferenceProvider());
+        registrar.registerReferenceProvider(PlatformPatterns.psiElement(QImportElement.class), new QImportReferenceProvider());
     }
 }